Refactor UserDir() method to Paths class
