+++ src/xp.runner/io/Paths.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Diagnostics;
 using System.Reflection;
 using System.Collections.Generic;
+using Xp.Runners;
 
 namespace Xp.Runners.IO
 {
@@ -40,7 +41,6 @@ namespace Xp.Runners.IO
             }
         }
 
-
         /// <summary>Locate a given file inside multiple base paths. Throw an exception if the file cannot
         /// be found. Similar to what is done when looking up program names in $ENV{PATH}.</summary>
         public static IEnumerable<string> Locate(IEnumerable<string> bases, IEnumerable<string> files)
@@ -62,6 +62,41 @@ namespace Xp.Runners.IO
         public static string Home()
         {
             return Environment.GetEnvironmentVariable("HOME") ?? Environment.GetFolderPath(Environment.SpecialFolder.Personal);
+        }
+
+        /// <summary>Returns whether the environment indicates this system conforms to
+        /// https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html</summary>
+        public static bool UseXDG()
+        {
+            foreach (string variable in Environment.GetEnvironmentVariables().Keys)
+            {
+                if (variable.StartsWith("XDG_")) return true;
+            }
+            return false;
+        }
+
+        /// <summary>Returns the user-specific config directory. Respects $HOME, XDG-compliant
+        /// systems and falls back to using %APPDATA%</summary>
+        public static string UserDir(string name)
+        {
+            if (UseXDG())
+            {
+                return Compose(
+                    Environment.GetEnvironmentVariable("XDG_CONFIG_HOME") ?? Compose(Home(), ".config"),
+                    name
+                );
+            }
+
+            var home = Environment.GetEnvironmentVariable("HOME");
+            if (String.IsNullOrEmpty(home))
+            {
+                return Compose(Environment.SpecialFolder.ApplicationData, Strings.UpperCaseFirst(name));
+            }
+            else
+            {
+                return Compose(home, "." + name);
+            }
+
         }
 
         /// <summary>Resolve a path. If the path is actually a shell link (.lnk file), this link's target path is used</summary>
